Remove unused rollup-plugin-livereload dependency
The 'rollup-plugin-livereload' dependency was removed from package.json as it's no longer in use. This helps to keep the codebase clean and the build process faster.

Corresponding lines related to this package are also removed from yarn.lock file.
Also, .idea was added to .gitignore to prevent accidentally checking in IDE-specific files.